Prevent Basic Authentication dialog
When an invalid_client error occours, the response will contain
a header WWW-Authenticate: Basic realm="OAuth". This causes the
browser to popup a dialog with basic authentication and the user
does not see the error at first. This change removes this header
when the error is being handled.
see: https://github.com/thephpleague/oauth2-server/issues/745